requirements.txt
bigquery==0.0.38
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dacktool==0.0.7
dbstream==0.1.21
dnspython==2.4.2
flask==3.0.0
google-api-core==2.15.0
google-api-python-client==1.7.11
google-auth==2.26.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.39.0
google-cloud-bigquery==3.15.0
google-cloud-bigquery-reservation==1.12.0
google-cloud-bigquery-storage==2.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-secret-manager==2.7.2
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
googleauthentication==0.0.17
grpc-google-iam-v1==0.12.7
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
libcst==1.1.0
MarkupSafe==2.1.3
motor==3.3.2
mypy-extensions==1.0.0
numpy==1.24.4
oauthlib==3.2.2
packaging==23.2
pandas==1.3.4
proto-plus==1.23.0
protobuf==4.25.1
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
shapely==2.0.2
six==1.16.0
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
uritemplate==3.0.1
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
